# AdminPanel/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CalendarRestApi",
    "version": "1.0"
  },
  "paths": {
    "/Calendar/{calendarName}/HolidaysWithPeriodDate/{startDate}/{endDate}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calendar/{calendarName}/IsWorkingDay/{date}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calendar/{calendarName}/NextWorkingDate/{date}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calendar/{calendarName}/GetNextWorkingsDate/{date}/{step}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "step",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calendar/{calendarName}/StatusDate/{date}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calendar/{calendarName}/WorkingDayCount/{startDate}/{endDate}": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Calendar/{calendarName}/Events/Add": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Events/{eventDate}/Get": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DateEvent"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateEvent"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DateEvent"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Events/{eventDate}/Delete": {
      "delete": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Events/Update": {
      "put": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Add": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Get": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calendar"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/GetAll": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "operationId": "GetAllCalendars",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Calendar"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Calendar"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Calendar"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Remove": {
      "delete": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/{date}/Weekends/Get": {
      "get": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayOfWeek"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayOfWeek"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DayOfWeek"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Weekends/Modify": {
      "put": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    },
    "/Calendar/{calendarName}/Weekends/Add": {
      "post": {
        "tags": [
          "Calendar"
        ],
        "parameters": [
          {
            "name": "calendarName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DayOfWeek"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [
              "calendar.api"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Calendar": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "weekend": {
            "$ref": "#/components/schemas/DayOfWeekListAffectedByDateCollection"
          },
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateEvent"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DateEvent": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "description": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "isHoliday": {
            "type": "boolean",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DateTimeDayOfWeekListKeyValuePair": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "format": "date-time"
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayOfWeek"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [
          "Sunday",
          "Monday",
          "Tuesday",
          "Wednesday",
          "Thursday",
          "Friday",
          "Saturday"
        ],
        "type": "string"
      },
      "DayOfWeekListAffectedByDateCollection": {
        "type": "object",
        "properties": {
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DateTimeDayOfWeekListKeyValuePair"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EventDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isHoliday": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://si-lab-idp.etadbir.com/connect/authorize",
            "tokenUrl": "https://si-lab-idp.etadbir.com/connect/token",
            "scopes": {
              "calendar.api.scope": ""
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "ClientKey": []
    }
  ]
}